requirements.txt
Django==1.7.1
Pillow==2.6.1
django-model-utils==2.2
djangorestframework==3.0.4
psycopg2==2.5.4
pytils==0.3
